refactor: Split EmailPasswordInteractionPolicy
