+++ config/presets/identity/email-password-interaction-policy.json
@@ -215,10 +215,7 @@
     {
       "comment": "Sets up the email password interaction policy",
       "@id": "urn:solid-server:email-password-interaction:EmailPasswordInteractionPolicy",
-      "@type": "EmailPasswordInteractionPolicyHttpHandler",
-      "EmailPasswordInteractionPolicyHttpHandler:_args_interactionHttpHandler": {
-        "@id": "urn:solid-server:email-password-interaction:InteractionHttpHandler"
-      }
+      "@type": "EmailPasswordInteractionPolicy"
     }
   ]
 }
