--- test/unit/identity/interaction/IdpInteractionPolicyHttpHandler.test.ts
@@ -1,9 +1,0 @@
-import {
-  IdpInteractionPolicyHttpHandler,
-} from '../../../../src/identity/interaction/IdpInteractionPolicyHttpHandler';
-
-describe('IdpInteractionHttpHandler', (): void => {
-  it('satisfies a trivial use case.', async(): Promise<void> => {
-    expect(IdpInteractionPolicyHttpHandler).toBeDefined();
-  });
-});
