--- test/unit/identity/interaction/email-password/EmailPasswordInteractionPolicyHttpHandler.test.ts
@@ -1,9 +1,0 @@
-import {
-  EmailPasswordInteractionPolicyHttpHandler,
-} from '../../../../../src/identity/interaction/email-password/EmailPasswordInteractionPolicyHttpHandler';
-
-describe('EmailPasswordInteractionPolicyHttpHandler', (): void => {
-  it('satisfies a trivial use case.', async(): Promise<void> => {
-    expect(EmailPasswordInteractionPolicyHttpHandler).toBeDefined();
-  });
-});
